requirements.txt
flask>=3.1.0
numpy==1.26.2
pillow==10.1.0
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1